adding example to readme, simplifying
simplifying adoption of clacks_rpc by not requiring the user to use an RPC server, as we have no need for a custom server class.

Handled nested getattr__ calls for safety, but magic methods in this interface can still be accessed normally.

As clacks is not meant to be secure, and RPC is not secure anyway, this is OK, as it is the developer's discretion to use these methods correctly.